Rename *LTT components
