+++ src/routes/share/+page.svelte
@@ -1,22 +1,22 @@
 <script lang="ts">
-  import HealthLinkLTT from '$lib/HealthLinkLTT.svelte';
-  import AddFileLTT from '$lib/AddFileLTT.svelte';
+    import HealthLink from '$lib/HealthLink.svelte';
+    import AddFile from '$lib/AddFile.svelte';
+  
+    let shlReady = false;
 
-  let shlReady = false;
+    function updateReady(ready: boolean) {
+      shlReady = ready;
+    }
+  </script>
 
-  function updateReady(ready: boolean) {
-    shlReady = ready;
-  }
-</script>
+  <svelte:head>
+    <title>Let's Talk Tech - Share</title> 
+  </svelte:head>
+  
+  <AddFile
+    on:shl-ready={ ({ detail }) => { updateReady(detail) } }
+  />
 
-<svelte:head>
-  <title>Let's Talk Tech - Share</title> 
-</svelte:head>
-
-<AddFileLTT
-  on:shl-ready={ ({ detail }) => { updateReady(detail) } }
-/>
-
-{#if shlReady}
-  <HealthLinkLTT/>
-{/if}
+  {#if shlReady}
+    <HealthLink/>
+  {/if}
